Se agrago la funcionalidad de tener como emisor el celular para la lectura de otro dispositivo
