+++ android/app/src/main/AndroidManifest.xml
@@ -1,4 +1,11 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
+    <!-- necesarios para  el funcionamiento de nfc -->
+    <uses-permission android:name="android.permission.INTERNET"/>
+    <uses-permission android:name="android.permission.NFC" />
+    <uses-feature android:name="android.hardware.nfc" android:required="true" />
+    <uses-feature android:name="android.hardware.nfc.hce" android:required="true" />
+    <uses-permission android:name="android.permission.VIBRATE" />
+
     <application
         android:label="nfc_demo"
         android:name="${applicationName}"
@@ -24,13 +31,26 @@
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
+   
+        <!-- flutter_nfc_hce service add begin -->
+        <service android:name="com.novice.flutter_nfc_hce.KHostApduService"
+                 android:exported="true"
+                 android:enabled="true"
+                 android:permission="android.permission.BIND_NFC_SERVICE">
+            <intent-filter>
+                <action android:name="android.nfc.cardemulation.action.HOST_APDU_SERVICE"/>
+                <category android:name="android.intent.category.DEFAULT"/>
+            </intent-filter>
+            <meta-data android:name="android.nfc.cardemulation.host_apdu_service"
+                       android:resource="@xml/apduservice"/>
+        </service>
+        <!-- flutter_nfc_hce service add end -->
+        
         <!-- Don't delete the meta-data below.
              This is used by the Flutter tool to generate GeneratedPluginRegistrant.java -->
         <meta-data
             android:name="flutterEmbedding"
             android:value="2" />
     </application>
-    <!-- necesarios para  el funcionamiento de nfc -->
-    <uses-permission android:name="android.permission.INTERNET"/>
-     <uses-permission android:name="android.permission.NFC"/>
+
 </manifest>
