+++ android/app/src/main/res/xml/apduservice.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<host-apdu-service xmlns:android="http://schemas.android.com/apk/res/android"
+                   android:description="@string/servicedesc" android:requireDeviceScreenOn="false" android:requireDeviceUnlock="false">
+    <aid-group android:description="@string/aiddescription"  android:category="other" >
+        <aid-filter android:name="D2760000850101"/>
+    </aid-group>
+</host-apdu-service>
